+++ activerecord/lib/active_record/connection_adapters/postgresql/schema_statements.rb
@@ -387,7 +387,7 @@ module ActiveRecord
         end
 
         # Maps logical Rails types to PostgreSQL-specific data types.
-        def type_to_sql(type, limit = nil, precision = nil, scale = nil)
+        def type_to_sql(type, limit = nil, precision = nil, scale = nil, fixed = nil)
           case type.to_s
           when 'binary'
             # PostgreSQL doesn't support limits on binary (bytea) columns.